refactor: restructure auth module and update related files
