+++ src/modules/user/user.controller.ts
@@ -6,10 +6,10 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { UserService } from './user.service';
-import { JwtAuthGuard } from '../auth/auth.guard';
-import { RoleGuard } from '../auth/role.guard';
-import { Roles } from '../auth/role.decorator';
-import { UserRole } from './user.entity';
+import { JwtAuthGuard } from '../auth/guards/auth.guard';
+import { RoleGuard } from '../auth/guards/role.guard';
+import { Roles } from '../auth/decorators/role.decorator';
+import { UserRole } from '../user/user.entity';
 
 @Controller('users')
 @UseGuards(JwtAuthGuard, RoleGuard)
